feat(examples): use @refinedev/rest in data-provider-strapi-v4 example
