+++ examples/data-provider-strapi-v4/package.json
@@ -27,10 +27,12 @@
     "@refinedev/cli": "^2.16.48",
     "@refinedev/core": "^5.0.1",
     "@refinedev/react-router": "^2.0.0",
+    "@refinedev/rest": "^1.0.0",
     "@refinedev/strapi-v4": "^7.0.0",
     "@uiw/react-md-editor": "^4.0.8",
     "antd": "^5.23.0",
     "axios": "^1.11.0",
+    "ky": "^1.10.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router": "^7.0.2"
